🗺️ Atlas: [parser structure] Extract pest parser into core module
Untangled the circular dependencies in `src/parser`. `grammar.rs` previously contained both the raw Pest struct (`GlossaParser` / `Rule`) and the high-level `parse` function, creating a dependency web where `statements` and `expressions` depended on `grammar`, while `grammar` (indirectly via `mod.rs`) depended on them. By extracting `GlossaParser` and `Rule` into `src/parser/core.rs`, the dependency graph is now a clean tree. Also suppressed unused variable warning in `main.rs`.
