+++ src/parser/core.rs
@@ -1,0 +1,15 @@
+use pest_derive::Parser;
+
+/// The PEG-based parser for ΓΛΩΣΣΑ.
+///
+/// This struct is automatically generated by the [`pest_derive`] crate based
+/// on the `grammar.pest` file. It provides the low-level string-to-token matching
+/// logic.
+///
+/// **Note:** End users generally do not interact with this struct directly.
+/// The `parse` function in the `grammar` module provides a convenient wrapper.
+#[derive(Parser)]
+#[grammar = "parser/grammar.pest"]
+#[doc(hidden)]
+#[allow(missing_docs)]
+pub struct GlossaParser;
