+++ src/parser/grammar.rs
@@ -54,30 +54,7 @@
 #![allow(missing_docs)]
 
 use pest::Parser;
-use pest_derive::Parser;
-
-/// The PEG-based parser for ΓΛΩΣΣΑ.
-///
-/// This struct is automatically generated by the [`pest_derive`] crate based
-/// on the `grammar.pest` file. It provides the low-level string-to-token matching
-/// logic.
-///
-/// **Note:** End users generally do not interact with this struct directly.
-/// The [`parse`] function in this module provides a convenient wrapper.
-///
-/// # Examples
-///
-/// ```
-/// use glossa::parser::grammar_parse_public as parse;
-///
-/// let pairs = parse("«χαῖρε» λέγε.");
-/// assert!(pairs.is_ok());
-/// ```
-#[derive(Parser)]
-#[grammar = "parser/grammar.pest"]
-#[doc(hidden)]
-#[allow(missing_docs)]
-pub struct GlossaParser;
+pub use crate::parser::core::{GlossaParser, Rule};
 
 /// Parse a ΓΛΩΣΣΑ source string into a pest parse tree (Concrete Syntax Tree)
 ///
