+++ src/parser/mod.rs
@@ -23,6 +23,7 @@
 //! recursive descent parsing phase.
 
 pub(crate) mod common;
+pub(crate) mod core;
 pub(crate) mod declarations;
 pub(crate) mod expressions;
 pub(crate) mod grammar;
@@ -30,7 +31,8 @@ pub(crate) mod numerals;
 pub(crate) mod recursion;
 pub(crate) mod statements;
 
-use self::grammar::{Rule, parse as grammar_parse};
+use self::core::Rule;
+use self::grammar::parse as grammar_parse;
 use crate::ast::*;
 use crate::errors::GlossaError;
 use pest::iterators::Pair;
@@ -306,7 +308,7 @@ mod tests_coverage {
 
     #[test]
     fn test_parse_unexpected_error_handling() {
-        use crate::parser::grammar::{GlossaParser, Rule};
+        use crate::parser::core::{GlossaParser, Rule};
         use pest::Parser;
 
         let mut pairs = GlossaParser::parse(Rule::string_literal, "«a»").unwrap();
@@ -318,7 +320,7 @@ mod tests_coverage {
 
     #[test]
     fn test_parse_empty_statement_handling() {
-        use crate::parser::grammar::{GlossaParser, Rule};
+        use crate::parser::core::{GlossaParser, Rule};
         use pest::Parser;
 
         let mut pairs = GlossaParser::parse(Rule::period, ".").unwrap();
