+++ src/parser/statements.rs
@@ -4,7 +4,7 @@
 use crate::ast::*;
 use crate::parser::common::ParseError;
 use crate::parser::expressions::build_expression;
-use crate::parser::grammar::Rule;
+use crate::parser::core::Rule;
 use pest::iterators::Pair;
 
 pub(crate) fn build_regular_statement(
@@ -70,7 +70,7 @@ fn parse_statement_end(pair: Pair<'_, Rule>) -> (bool, bool) {
 #[cfg(test)]
 mod tests_coverage {
     use super::*;
-    use crate::parser::grammar::{GlossaParser, Rule};
+    use crate::parser::core::{GlossaParser, Rule};
     use pest::Parser;
 
     #[test]
